+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/AbstractSqlObjectFactory.java
@@ -47,6 +47,8 @@ abstract class AbstractSqlObjectFactory implements ExtensionFactory {
         builder.addMethodHandler(JdbiClassUtils.TOSTRING_METHOD, toStringHandler);
         builder.addMethodHandler(GET_HANDLE_METHOD, GET_HANDLE_HANDLER);
         builder.addMethodHandler(WITH_HANDLE_METHOD, WITH_HANDLE_HANDLER);
+
+        new DefinitionsFactory(extensionType).configureDefinitions(builder);
     }
 
     @Override