+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- New Feature: annotate types, methods, or fields as `@Definition` to define computed constants
 - fix edge condition when calling `Connection#commit()` threw an Exception, we called commit() again. Now the
   code explicitly calls `Connection#rollback()`. Fixes #2595
 - FieldMapper: skip static fields (#2607, reported by @mvysny)